requirements.txt
selenium==4.12.0
pytest==7.4.2
black==25.1.0
allure-pytest==2.13.5
pytest-selenium==4.1.0